Tracing Framework/C/trace-modeling-oss-fuzz-c/requirements.txt
numpy
pandas
tqdm
tree-sitter
tree-sitter-languages
PyYAML
jsonlines
lxml
